Replace usage of Console.MoveBufferArea on Unix
Console.MoveBufferArea is a Windows-only API. It's a performance
optimization as regular console output is extremely slow. Fortunately,
this problem doesn't even seem to exist (well, at least on macOS,
because that's where I was testing) so we can just get away with doing
the naive output when not running on Windows.