Improve performance
For smaller repos, the original version was plenty fast. But with larger repos
staging was painfully slow. This speeds it up quite a bit, but ideally we'd
make it much faster still. I'm a bit at a lass what can be done; one option
might be to stop using LibGit2Sharp and just shelling out for all operations
as CLI git seems to be quicker. However, this isn't super straight forward as
git diff doesn't support showing untracked files as being added, which we
most certainly want. This would also require parsing patch files directly,
but that's not a big deal IMHO.